+++ lib/dom/spinner.dart
@@ -1,0 +1,14 @@
+part of sg2o;
+
+/// Simple element with a spinning circle inside.
+class Spinner extends HtmlElement {
+
+    factory Spinner() {
+        Element spinner = new Element.tag('i')
+            ..classes.add(classFontAwesome)
+            ..classes.add(classFASpinner)
+            ..classes.add(classFASpin)
+        ;
+        return spinner;
+    }
+}