+++ lib/styles/tooltip.dart
@@ -8,6 +8,29 @@ class Tooltip extends Styles {
     /// Adds all border styles to the DOM.
     void addStyles() {
 
+        // This is only for the profile tooltip when hovering over the avatar on a giveaway.
+        styleSheet.insertRule(
+            '''
+            .$classProfileTooltip {
+                width: 550px;
+                padding: 5px;
+            }
+            ''', 0);
+
+        styleSheet.insertRule(
+            '''
+            #$classProfileTooltipContainer {
+                width: 560px;
+                height: 195px;
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                display:none;
+            }
+            ''', 0);
+
+        // ###############################################
+        // Following the stylesheet for the tooltips for icons on a giveaway.
         styleSheet.insertRule(
             '''
             .$classTooltip {